+++ IntegrationBus/include/ib/capi/InterfaceIdentifiers.h
@@ -42,4 +42,7 @@ typedef int32_t ib_InterfaceIdentifier;
 #define ib_InterfaceIdentifier_RpcExchangeFormat           ((ib_InterfaceIdentifier)6001001)
 #define ib_InterfaceIdentifier_RpcDiscoveryResult          ((ib_InterfaceIdentifier)6002001)
 
+// Participant
+#define ib_InterfaceIdentifier_ParticipantStatus           ((ib_InterfaceIdentifier)7001001)
+
 IB_END_DECLS